Issue #402: Move WebView.deleteData to ext fn.
